+++ src/Plugin.Fingerprint.Android/Contract/IAuthenticationFailedListener.cs
@@ -1,8 +1,0 @@
-namespace Plugin.Fingerprint.Contract
-{
-    public interface IAuthenticationFailedListener
-    {
-        void OnFailedTry();
-        void OnHelp(FingerprintAuthenticationHelp help, string nativeHelpText);
-    }
-}