+++ src/Plugin.Fingerprint/Platforms/Android/CrossFingerprint.Android.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using Android.App;
+
+namespace Plugin.Fingerprint
+{
+    public partial class CrossFingerprint
+    {
+        private static Func<Activity> _activityResolver;
+
+        public static Activity CurrentActivity => GetCurrentActivity();
+        
+        public static void SetCurrentActivityResolver(Func<Activity> activityResolver)
+        {
+            _activityResolver = activityResolver;
+        }
+
+        private static Activity GetCurrentActivity()
+        {
+            if (_activityResolver == null)
+                throw new InvalidOperationException("Resolver for the current activity is not set. Call Fingerprint.SetCurrentActivityResolver somewhere in your startup code.");
+
+            return _activityResolver();
+        }
+    }
+}