+++ src/Sample.Android/MainActivity.cs
@@ -1,0 +1,23 @@
+﻿using Android.App;
+using Android.Content.PM;
+using Android.OS;
+using Plugin.Fingerprint;
+
+namespace Sample.Droid
+{
+    [Activity(Label = "Sample", Icon = "@mipmap/icon", Theme = "@style/MainTheme", MainLauncher = true, ConfigurationChanges = ConfigChanges.ScreenSize | ConfigChanges.Orientation)]
+    public class MainActivity : global::Xamarin.Forms.Platform.Android.FormsAppCompatActivity
+    {
+        protected override void OnCreate(Bundle savedInstanceState)
+        {
+            CrossFingerprint.SetCurrentActivityResolver(() => this);
+
+            TabLayoutResource = Resource.Layout.Tabbar;
+            ToolbarResource = Resource.Layout.Toolbar;
+
+            base.OnCreate(savedInstanceState);
+            global::Xamarin.Forms.Forms.Init(this, savedInstanceState);
+            LoadApplication(new App());
+        }
+    }
+}